Mark nodes as USED when deleting autohold
Marking the nodes as USED will allow nodepool to delete them.
If we are unsuccessful in marking any of the held nodes as used,
we simply log the error and try again at some future point until
all nodes are eventually marked, allowing the hold request to be
deleted.

+++ zuul/zk.py
@@ -583,12 +583,58 @@ class ZooKeeper(object):
             path = self.HOLD_REQUEST_ROOT + "/" + hold_request.id
             self.client.set(path, hold_request.serialize())
 
+    def _markHeldNodesAsUsed(self, hold_request):
+        '''
+        :returns: True if all nodes marked USED, False otherwise.
+        '''
+        failure = False
+        for node_id in hold_request.nodes:
+            node = self.getNode(node_id)
+            if not node or node['state'] == zuul.model.STATE_USED:
+                continue
+
+            node['state'] = zuul.model.STATE_USED
+
+            name = None
+            label = None
+            if 'name' in node:
+                name = node['name']
+            if 'label' in node:
+                label = node['label']
+
+            node_obj = zuul.model.Node(name, label)
+            node_obj.updateFromDict(node)
+
+            try:
+                self.lockNode(node_obj, blocking=False)
+                self.storeNode(node_obj)
+            except Exception:
+                self.log.exception("Cannot change HELD node state to USED "
+                                   "for node %s in request %s",
+                                   node_obj.id, hold_request.id)
+                failure = True
+            finally:
+                try:
+                    if node_obj.lock:
+                        self.unlockNode(node_obj)
+                except Exception:
+                    self.log.exception(
+                        "Failed to unlock HELD node %s for request %s",
+                        node_obj.id, hold_request.id)
+
+        return not failure
+
     def deleteHoldRequest(self, hold_request):
         '''
         Delete a hold request.
 
         :param HoldRequest hold_request: Object representing the hold request.
         '''
+        if not self._markHeldNodesAsUsed(hold_request):
+            self.log.info("Unable to delete hold request %s because "
+                          "not all nodes marked as USED.", hold_request.id)
+            return
+
         path = self.HOLD_REQUEST_ROOT + "/" + hold_request.id
         try:
             self.client.delete(path, recursive=True)
